Fix unsigned underflow in auto-select-speed timing diff
vp8/encoder/onyx_int.h declares cpi->avg_pick_mode_time and
cpi->avg_encode_time as signed int. govpx mirrored both as uint64. On
the first inter frame after a key frame, libvpx's vp8_auto_select_speed
evaluates `(avg_encode_time - avg_pick_mode_time) < ms_for_compress`
where avg_encode_time is still 0 (KFs are excluded from the IIR update)
and avg_pick_mode_time is positive. As signed int this is a small
negative number and the guard succeeds, keeping the auto-Speed at 4 (or
in the slow ramp). As uint64 it underflowed to ~MAX_UINT64, the guard
failed, and govpx took the speed-bump branch (autoSpeed += 4) on frame
1. The Speed=8 path then enabled the fast LF picker, while libvpx kept
auto_filter=1 + the full LF picker, which produces a different
filter_level (e.g. 11 vs 5) on inter frames and propagates as larger
interframe sizes. With the types switched to int, the difference is
signed-negative on frame 1 and govpx tracks libvpx's auto-Speed
trajectory.

Diagnosis flow: TestOracleLFTrialDiag (GOVPX_DEBUG=1) shows govpx using
fast-picker rows (seed/up) and libvpx using full-picker rows on inter
frames at cpu_used=8 RT. After patch govpx emits the same "full" row
shape and frame 2 picks LF=8 to match libvpx. 720p RT cpu8 1500kbps
bench drops avg interframe from 10.35 KiB to 9.89 KiB (libvpx 8.41
KiB).

Inter-mode distribution baseline ratchets: l1_pp shrinks across all
fixtures (rt-cpu8-128x128-bench-noise 3.12->1.79, rt-cpu8-128x128-
panning 1.79->0.45, rt-cpu8-256x256-bench-noise 3.46->1.45) and
eob_sum_ratio converges toward 1.0.

+++ encoder.go
@@ -245,9 +245,9 @@ type VP8Encoder struct {
 	// the timers IIR-update (7*avg + duration)>>3. Cold start: Speed=4,
 	// timers=0. Govpx tracks the same state and feeds e.autoSpeed to
 	// libvpxCPUUsed for realtime+positive-cpu_used branches.
-	autoSpeed             int    // current adaptive Speed (libvpx cpi->Speed)
-	avgPickModeTime       uint64 // microseconds (libvpx cpi->avg_pick_mode_time)
-	avgEncodeTime         uint64 // microseconds (libvpx cpi->avg_encode_time)
+	autoSpeed             int // current adaptive Speed (libvpx cpi->Speed)
+	avgPickModeTime       int // microseconds (libvpx cpi->avg_pick_mode_time, signed int per onyx_int.h)
+	avgEncodeTime         int // microseconds (libvpx cpi->avg_encode_time, signed int per onyx_int.h)
 	autoSpeedFrameStartNS int64
 
 	// libvpx vp8/encoder/onyx_if.c forced-key bookkeeping. this_key_frame_forced
@@ -2974,12 +2974,21 @@ func (e *VP8Encoder) libvpxAutoSelectSpeed() {
 	}
 	msForCompress := 1000000 / fps
 	msForCompress = msForCompress * (16 - cpuUsed) / 16
-	if e.avgPickModeTime < uint64(msForCompress) &&
-		(e.avgEncodeTime-e.avgPickModeTime) < uint64(msForCompress) {
+	// Note: avgPickModeTime and avgEncodeTime are signed int to mirror
+	// libvpx's signed-int avg_pick_mode_time / avg_encode_time
+	// (vp8/encoder/onyx_int.h:455-456). The (avgEncodeTime - avgPickModeTime)
+	// difference is signed-negative on the first inter frame after a key
+	// frame (avg_encode_time is skipped for KFs while avg_pick_mode_time is
+	// updated, so EncodeTime=0 < PickModeTime). Using uint64 here would let
+	// the subtraction underflow to a huge unsigned value, fail the
+	// "< msForCompress" guard, and force the bump branch (Speed += 4) on
+	// frame 1, breaking the auto-select trajectory libvpx follows.
+	if e.avgPickModeTime < msForCompress &&
+		(e.avgEncodeTime-e.avgPickModeTime) < msForCompress {
 		if e.avgPickModeTime == 0 {
 			e.autoSpeed = 4
 		} else {
-			if msForCompress*100 < int(e.avgEncodeTime)*95 {
+			if msForCompress*100 < e.avgEncodeTime*95 {
 				e.autoSpeed += 2
 				e.avgPickModeTime = 0
 				e.avgEncodeTime = 0
@@ -2988,7 +2997,7 @@ func (e *VP8Encoder) libvpxAutoSelectSpeed() {
 				}
 			}
 			if e.autoSpeed >= 0 && e.autoSpeed < len(libvpxAutoSpeedThresh) &&
-				msForCompress*100 > int(e.avgEncodeTime)*libvpxAutoSpeedThresh[e.autoSpeed] {
+				msForCompress*100 > e.avgEncodeTime*libvpxAutoSpeedThresh[e.autoSpeed] {
 				e.autoSpeed -= 1
 				e.avgPickModeTime = 0
 				e.avgEncodeTime = 0
@@ -3020,7 +3029,7 @@ func (e *VP8Encoder) finishAutoSpeedTiming(isKeyFrame bool) {
 	if durationNS < 0 {
 		durationNS = 0
 	}
-	duration := uint64(durationNS / 1000) // microseconds
+	duration := int(durationNS / 1000) // microseconds
 	duration2 := duration / 2
 	if !isKeyFrame {
 		if e.avgEncodeTime == 0 {
